requirements.txt
py-pure-client==1.46.0
pytest-dependency==0.6.0
